+++ NEWS.md
@@ -12,6 +12,9 @@
 ### folio-spring-testing
 * [FOLSPRINGS-238](https://folio-org.atlassian.net/browse/FOLSPRINGS-238) Fix MinioContainerExtension image pull failure by switching to quay.io/minio/minio
 
+### folio-spring-kafka-filtering
+* [FOLSPRINGS-232](https://folio-org.atlassian.net/browse/FOLSPRINGS-232) Add tenant-aware Kafka message filtering support for Spring modules
+
 ## 10.0.0 2026-04-06
 
 **Breaking Changes:**